feat: add CariCategoryFilter and CariSupplierGrid components for enhanced supplier filtering
